PERF: GDCM CanRead function fail quickly for non-compliant DICOM files.
Some images take extremely long time to load because GDCMImageIO::CanReadFile
takes several minutes to fail.  This is due to GDCM's default behavior of trying
to read non-compliant dicom files that do not have the required DICM header.

Add more extensive testing about the structure of the file to determine
if it looks like a dicom file.  Previous testing only looked to see if the
files without preables had values of 2 or 8 as the first byles of the file,
but that resulted in many false positives.

This implementation looks at all the SOP Instances that start with 2 or 8
to ensure that the proper dicom structure is found.

This resolves #388.